feat: asset

+++ HotSource/src/main/java/hotsource/domain/Role.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class Role {
 	
-	private int role_id;
+	private Long role_id;
 	private String role_name;
 
 }
