+++ HotSource/src/main/java/hotsource/domain/SnsProvider.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class SnsProvider {
 	
-	private int sns_provider_id;
+	private Long sns_provider_id;
 	private String provider_name;
 
 }
